removed unsigned vs int  warning rules.cc

+++ src/rules.cc
@@ -123,7 +123,7 @@ void Rules::resolve_score()
     {
         if (p->type == rules::PLAYER)
         {
-            if (api_->game_state()->get_player_artefact() == p->id)
+            if (api_->game_state()->get_player_artefact() == (int)p->id)
                 p->score += POINTS_CONTROLE_ARTEFACT;
